08/20/2025
created new function for \alpha_{ji}^k, calculates this like the adjacency matrix but between resource nodes only(this is the r to r infection rate). Also for \beta_{ij}^{wk} I developed a function that caluclates it the same was as the alpha function(resource to pop node infection rate).